refactor(metadata): extract MetadataStore struct into store.rs
Move MetadataStore struct, constructors, and connection helpers
(with_conn, with_conn_mut, with_sqlite_conn, in_clause_placeholders)
from mod.rs into a new store.rs submodule. Fields and crate-internal
helpers promoted to pub(super) so sibling impl blocks keep working.

No behavior change. No SQL change. Public API unchanged — MetadataStore
is re-exported at the same path via `pub use store::MetadataStore`.
mod.rs: 3747 → 3632 lines; store.rs: 133 lines.

Phase 1 of 6 for the metadata/mod.rs split.

Refs #291
Refs #290
